refactor: remove unnecessary auto(un)boxing

+++ src/net/sourceforge/plantuml/FileWatcher.java
@@ -61,7 +61,7 @@ public class FileWatcher {
 	public boolean hasChanged() {
 		for (Map.Entry<File, Long> ent : modified2.entrySet()) {
 			final long nowModified = ent.getKey().lastModified();
-			if (ent.getValue().longValue() != nowModified) {
+			if (ent.getValue() != nowModified) {
 				return true;
 			}
 		}
